修复业务抛异常导致SchedulerManager.runner永远为true从而无法执行后续调度任务的问题

+++ lib/src/main/java/yanry/lib/android/entity/BezierEvaluator.java
@@ -8,15 +8,15 @@ import yanry.lib.java.util.MathUtil;
  * Created by rongyu.yan on 3/21/2017.
  */
 
-public class BezierEvaluator implements TypeEvaluator<Integer> {
-    private int[] medianValues;
+public class BezierEvaluator implements TypeEvaluator<Float> {
+    private float[] medianValues;
 
-    public BezierEvaluator(int... medianValues) {
+    public BezierEvaluator(float... medianValues) {
         this.medianValues = medianValues;
     }
 
     @Override
-    public Integer evaluate(float fraction, Integer startValue, Integer endValue) {
+    public Float evaluate(float fraction, Float startValue, Float endValue) {
         return MathUtil.bezierEvaluate(fraction, startValue, endValue, medianValues);
     }
 }
